# src/modules/documents/entities/document.entity.ts
import { Entity, Column, PrimaryGeneratedColumn } from 'typeorm';
import { ApiProperty } from '@nestjs/swagger';

@Entity()
export class Document {
  @ApiProperty({
    description: 'Unique document id that was generated by server',
    example: 1,
  })
  @PrimaryGeneratedColumn()
  id: number;

  @ApiProperty({
    description:
      'MUST BE ENCRYPTED WITH ASSOCIATED KEY (stored in `keys` field). file name (with extension), under which document is stored on server',
    example: 'AAAAB3NzaC1yc2EAAAADAQABA',
  })
  @Column()
  name: string;

  @ApiProperty({
    description:
      'MUST BE ENCRYPTED WITH USER PERSONAL KEY. Key used to encrypt document name and content',
    example: 'TODO: add example',
  })
  @Column()
  keys: string;
}